+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	- `Only(HashSet<T>)`
   - `Slice(IReadOnlyList<T>)`
   - `Vector3 SwizzleXYZ_XZY(Vector3)`
+  - `AsRectTransform(Transform)` and `GetRectTransform(Component)`
 - `GetComponentCached` and `TryGetComponentCached` now also extend `Component` instead of just `GameObject`
 
 ### Fixed